LayoutTests/inspector/dom-debugger/event-animation-frame-breakpoints-expected.txt

Tests for Event AnimationFrame breakpoints.


== Running test suite: DOMDebugger.Event.AnimationFrame
-- Running test case: DOMDebugger.Event.AnimationFrame.AddBreakpoint "requestAnimationFrame"
Adding "requestAnimationFrame" Event Breakpoint...
Firing "requestAnimationFrame" on window...
PASS: Should pause before event handler is run.
CALL STACK:
0: [F] handleWindow_requestAnimationFrame
ASYNC CALL STACK:
1: --- requestAnimationFrame ---
2: [F] trigger_requestAnimationFrame
3: [P] Global Code
-- Running test teardown.

-- Running test case: DOMDebugger.Event.AnimationFrame.AddDisabledBreakpoint "requestAnimationFrame"
Adding "requestAnimationFrame" Event Breakpoint...
Disabling "requestAnimationFrame" Event Breakpoint...
Firing "requestAnimationFrame" on window...
PASS: Should not pause for disabled breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.AnimationFrame.RemoveBreakpoint "requestAnimationFrame"
Adding "requestAnimationFrame" Event Breakpoint...
Removing "requestAnimationFrame" Event Breakpoint...
Firing "requestAnimationFrame" on window...
PASS: Should not pause for removed breakpoint.
-- Running test teardown.

-- Running test case: DOMDebugger.Event.AnimationFrame.RemoveDisabledBreakpoint "requestAnimationFrame"
Adding "requestAnimationFrame" Event Breakpoint...
Disabling "requestAnimationFrame" Event Breakpoint...
Removing "requestAnimationFrame" Event Breakpoint...
Firing "requestAnimationFrame" on window...
PASS: Should not pause for removed disabled breakpoint.
-- Running test teardown.
